Extract queue player related types into separate files
