+++ Sources/Player/Internal/QueuePlayer.swift
@@ -7,24 +7,6 @@
 import AVFoundation
 import DequeModule
 
-enum SeekKey: String {
-    case time
-}
-
-struct Seek: Equatable {
-    let time: CMTime
-    let toleranceBefore: CMTime
-    let toleranceAfter: CMTime
-    let isSmooth: Bool
-    let completionHandler: (Bool) -> Void
-
-    private let id = UUID()
-
-    static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.id == rhs.id
-    }
-}
-
 class QueuePlayer: AVQueuePlayer {
     static let notificationCenter = NotificationCenter()
 
@@ -150,44 +132,4 @@ class QueuePlayer: AVQueuePlayer {
         mediaSelectionCriteria[mediaCharacteristic] = criteria
         super.setMediaSelectionCriteria(criteria, forMediaCharacteristic: mediaCharacteristic)
     }
-}
-
-extension AVQueuePlayer {
-    func replaceItems(with items: [AVPlayerItem]) {
-        guard self.items() != items else { return }
-
-        if let firstItem = items.first {
-            if firstItem !== self.items().first {
-                remove(firstItem)
-                // TODO: Workaround to fix incorrect recovery from failed MP3 playback (FB13650115). Remove when fixed.
-                if self.items().first?.error != nil {
-                    removeAllItems()
-                }
-                // End of workaround
-                replaceCurrentItem(with: firstItem)
-            }
-            removeAll(from: 1)
-            if items.count > 1 {
-                append(Array(items.suffix(from: 1)))
-            }
-        }
-        else {
-            removeAllItems()
-        }
-    }
-
-    private func removeAll(from index: Int) {
-        assert(index > 0, "The current item must not be removed")
-        guard items().count > index else { return }
-        items().suffix(from: index).forEach { remove($0) }
-    }
-
-    private func append(_ items: [AVPlayerItem]) {
-        items.forEach { insert($0, after: nil) }
-    }
-}
-
-extension Notification.Name {
-    static let willSeek = Notification.Name("QueuePlayerWillSeekNotification")
-    static let didSeek = Notification.Name("QueuePlayerDidSeekNotification")
 }
